Add integration and unit tests for TaskController and TaskService
Comprehensive integration and unit tests have been added for the TaskController and TaskService classes to ensure better test coverage and reliability. Minor adjustments have also been made to various project files to support these changes and improve overall project maintainability.

+++ src/main/java/com/javarush/kostenko/controller/TaskController.java
@@ -72,7 +72,7 @@ public class TaskController {
      */
     @PutMapping(value = "/tasks/edit/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
-    public ResponseEntity<Void> updateTask(@PathVariable Long id, @RequestBody Task taskUpdate) {
+    public ResponseEntity<Void> updateTask(@PathVariable Integer id, @RequestBody Task taskUpdate) {
         log.info("Updating task with id: {}", id);
         Optional<Task> existingTaskOptional = taskService.findById(id);
 
@@ -98,7 +98,7 @@ public class TaskController {
      */
     @DeleteMapping("/tasks/{id}")
     @ResponseBody
-    public ResponseEntity<Void> deleteTask(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteTask(@PathVariable Integer id) {
         log.info("Deleting task with id: {}", id);
         if (taskService.findById(id).isPresent()) {
             taskService.deleteById(id);
